feat: make phenyl-rest-api get session client from central-state
