+++ modules/central-state/src/index.ts
@@ -1,2 +1,3 @@
 export { PhenylEntityClient } from "./entity-client";
 export { PhenylEntityClientOptions } from "./entity-client";
+export { PhenylSessionClient } from "./session-client";
